Refactor YCSB-C database operations to use uintptr_t for value handling and implement AsyncScan functionality

+++ tests/YCSB-C/include/db/clevel_hash_db.h
@@ -23,11 +23,11 @@ public:
   //            std::vector<KVPair> &values);
   // int Delete(const std::string &table, const std::string &key);
 
-  int ReadInternal(uint64_t key, uint64_t &value) override;
+  int ReadInternal(uint64_t key, uintptr_t &value) override;
 
-  int UpdateInternal(uint64_t key, uint64_t value) override;
+  int UpdateInternal(uint64_t key, uintptr_t value) override;
 
-  int InsertInternal(uint64_t key, uint64_t value) override;
+  int InsertInternal(uint64_t key, uintptr_t value) override;
 
   int DeleteInternal(uint64_t key) override;
